@@ -56,20 +94,19 @@ configure(allprojects) { project ->
 			 "-Xlint:-deprecation", "-Xlint:-unchecked"]
 
 	compileJava {
-		sourceCompatibility = JavaVersion.VERSION_15
-		targetCompatibility = JavaVersion.VERSION_15
+		sourceCompatibility = JavaVersion.VERSION_17
+		targetCompatibility = JavaVersion.VERSION_17
 		options.encoding = "UTF-8"
 	}
 
 	compileTestJava {
-		sourceCompatibility = JavaVersion.VERSION_15
-		targetCompatibility = JavaVersion.VERSION_15
+		sourceCompatibility = JavaVersion.VERSION_17
+		targetCompatibility = JavaVersion.VERSION_17
 		options.encoding = "UTF-8"
 		options.compilerArgs += "-parameters"
 	}
 
 	repositories {
-	  	jcenter()
 	    mavenLocal()
 	    mavenCentral()
 	}
