@@ -1,6 +1,12 @@
 package org.forfun.mmorpg.game.cross.service;
 
-import com.google.common.collect.Maps;
+import jforgame.codec.struct.StructMessageCodec;
+import jforgame.socket.client.SocketClient;
+import jforgame.socket.netty.support.client.TcpSocketClient;
+import jforgame.socket.share.HostAndPort;
+import jforgame.socket.share.IdSession;
+import jforgame.socket.share.SocketIoDispatcher;
+import jforgame.socket.share.SocketIoDispatcherAdapter;
 import org.apache.commons.codec.digest.Md5Crypt;
 import org.forfun.mmorpg.game.CrossConfig;
 import org.forfun.mmorpg.game.ServerConfig;
@@ -8,28 +14,27 @@ import org.forfun.mmorpg.game.base.GameContext;
 import org.forfun.mmorpg.game.cross.message.RpcReqServerLogin;
 import org.forfun.mmorpg.game.cross.message.RpcServerNode;
 import org.forfun.mmorpg.game.cross.router.BalanceStrategy;
-import org.forfun.mmorpg.game.cross.router.RobinBalanceStrategy;
+import org.forfun.mmorpg.game.cross.router.RoundBalanceStrategy;
 import org.forfun.mmorpg.game.logger.LoggerUtils;
-import org.forfun.mmorpg.net.HostPort;
-import org.forfun.mmorpg.net.client.RpcClientFactory;
-import org.forfun.mmorpg.net.dispatcher.IMessageDispatcher;
-import org.forfun.mmorpg.net.socket.IdSession;
-import org.forfun.mmorpg.protocol.codec.SerializerFactory;
+import org.forfun.mmorpg.game.util.JsonUtil;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.io.IOException;
+import java.util.ArrayList;
 import java.util.List;
+import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ConcurrentMap;
 import java.util.stream.Collectors;
 
 @Service
 public class RpcClientRouter {
 
-    private ConcurrentMap<Integer, IdSession> servers = Maps.newConcurrentMap();
+    private ConcurrentMap<Integer, IdSession> servers = new ConcurrentHashMap<>();
 
-    private ConcurrentMap<Integer, RpcServerNode> nodes = Maps.newConcurrentMap();
+    private ConcurrentMap<Integer, RpcServerNode> nodes = new ConcurrentHashMap<>();
 
-    private BalanceStrategy fightBalanceStrategy = new RobinBalanceStrategy();
+    private BalanceStrategy fightBalanceStrategy = new RoundBalanceStrategy();
 
     @Autowired
     private CrossConfig crossConfig;
@@ -50,19 +55,25 @@ public class RpcClientRouter {
         if (servers.containsKey(targetSid)) {
             return servers.get(targetSid);
         }
+        SocketIoDispatcher msgDispatcher = new SocketIoDispatcherAdapter() {
+            @Override
+            public void dispatch(IdSession session, Object message) {
+                System.err.println("收到消息<-- " + message.getClass().getSimpleName() + "=" + JsonUtil.object2String(message));
+            }
+            @Override
+            public void exceptionCaught(IdSession session, Throwable cause) {
+                cause.printStackTrace();
+            }
+        };
         synchronized (this) {
             if (!servers.containsKey(targetSid)) {
-                IMessageDispatcher messageDispatcher = GameContext.getMessageDispatcher();
-                SerializerFactory serializerFactory = GameContext.getMessageSerializer();
-                RpcClientFactory clientFactory = new RpcClientFactory(messageDispatcher, serializerFactory);
-                HostPort hostPort = new HostPort();
-                hostPort.setHost(crossConfig.getCenterIp());
-                hostPort.setPort(crossConfig.getCenterPort());
+                SocketClient socketClient = new TcpSocketClient(msgDispatcher, GameContext.getMessageFactory(), new StructMessageCodec(),
+                        HostAndPort.valueOf(crossConfig.getCenterIp(),crossConfig.getCenterPort()));
                 try {
-                    IdSession session = clientFactory.createSession(hostPort);
+                    IdSession session = socketClient.openSession();
                     RpcReqServerLogin reqServerLogin = buildServerLoginRequest();
-                    session.sendPacket(reqServerLogin);
-                } catch (Exception e) {
+                    session.send(reqServerLogin);
+                } catch (IOException e) {
                     LoggerUtils.error("", e);
                     return null;
                 }
@@ -75,19 +86,23 @@ public class RpcClientRouter {
         if (!servers.containsKey(targetSid)) {
             return false;
         }
-        return servers.get(targetSid).isValid();
+        return servers.get(targetSid) != null;
     }
 
     public List<RpcServerNode> listRpcNodes() {
-        return nodes.values().stream().collect(Collectors.toList());
+        return new ArrayList<>(nodes.values());
+    }
+
+    public List<RpcServerNode> listRpcNodes(int serverType) {
+        return nodes.values().stream().filter(rpcServerNode -> rpcServerNode.getType() == serverType).collect(Collectors.toList());
     }
 
     public void registerSession(int targetSid, IdSession session) {
         servers.put(targetSid, session);
     }
 
-    public void registerRpcNode(int serverId, RpcServerNode node) {
-        nodes.put(serverId, node);
+    public void registerRpcNode(RpcServerNode node) {
+        nodes.put(node.getSid(), node);
     }
 
     private RpcReqServerLogin buildServerLoginRequest() {
