@@ -1,12 +1,12 @@
 package org.forfun.mmorpg.game.database.config.domain;
 
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.Id;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.Id;
 
 import lombok.Getter;
 
-@Entity
+@Entity(name = "configskill")
 @Getter
 public class ConfigSkill {
 
