@@ -1,11 +1,15 @@
 package org.forfun.mmorpg.game.database.converter;
 
-import org.apache.commons.lang3.StringUtils;
-import org.codehaus.jackson.map.ObjectMapper;
-import org.codehaus.jackson.map.type.TypeFactory;
-import org.codehaus.jackson.type.JavaType;
+import com.fasterxml.jackson.annotation.JsonAutoDetect;
+import com.fasterxml.jackson.annotation.JsonInclude;
+import com.fasterxml.jackson.annotation.PropertyAccessor;
+import com.fasterxml.jackson.databind.DeserializationFeature;
+import com.fasterxml.jackson.databind.JavaType;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.databind.type.TypeFactory;
+import jakarta.persistence.AttributeConverter;
+import org.springframework.util.StringUtils;
 
-import javax.persistence.AttributeConverter;
 import java.io.IOException;
 
 public class JpaObjectConverter implements AttributeConverter<Object, String> {
@@ -18,6 +22,11 @@ public class JpaObjectConverter implements AttributeConverter<Object, String> {
         MAPPER = new ObjectMapper();
         // 这里要写入类和属性的完整类型信息，反序列化即使使用Object.class也能解析
         MAPPER.enableDefaultTyping(ObjectMapper.DefaultTyping.NON_FINAL);
+        MAPPER.setSerializationInclusion(JsonInclude.Include.NON_NULL);
+        MAPPER.disable(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES);
+        MAPPER.setVisibility(PropertyAccessor.ALL, JsonAutoDetect.Visibility.NONE);
+        MAPPER.setVisibility(PropertyAccessor.FIELD, JsonAutoDetect.Visibility.ANY);
+        MAPPER.configure(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES, false);
     }
 
     @Override
@@ -25,7 +34,7 @@ public class JpaObjectConverter implements AttributeConverter<Object, String> {
         try {
             return MAPPER.writeValueAsString(attribute);
         } catch (Exception e) {
-            return null;
+            throw new RuntimeException(e);
         }
     }
 
@@ -38,9 +47,8 @@ public class JpaObjectConverter implements AttributeConverter<Object, String> {
             JavaType type = typeFactory.constructType(Object.class);
             return MAPPER.readValue(dbData, type);
         } catch (IOException e) {
-            e.printStackTrace();
+            throw new RuntimeException(e);
         }
-        return null;
     }
 
 }
