@@ -1,10 +1,10 @@
 package org.forfun.mmorpg.client;
 
+import jforgame.socket.share.IdSession;
 import org.forfun.mmorpg.game.gm.message.ReqGmCommand;
 import org.forfun.mmorpg.game.player.message.ReqAccountLogin;
 import org.forfun.mmorpg.game.player.message.ReqCreateNewPlayer;
 import org.forfun.mmorpg.game.player.message.ReqSelectPlayer;
-import org.forfun.mmorpg.net.socket.IdSession;
 
 public class ClientRobot {
 
@@ -17,25 +17,25 @@ public class ClientRobot {
     public void sendGm() {
         ReqGmCommand req = new ReqGmCommand();
         req.setParams("level 99");
-        session.sendPacket(req);
+        session.send(req);
     }
 
     public void createNew() {
         ReqCreateNewPlayer req = new ReqCreateNewPlayer();
         req.setName("Happy");
-        session.sendPacket(req);
+        session.send(req);
     }
 
     public void login() {
         ReqAccountLogin req = new ReqAccountLogin();
         req.setPassword("forfun");
         req.setAccountId(123L);
-        session.sendPacket(req);
+        session.send(req);
     }
 
     public void selectedPlayer(long playerId) {
         ReqSelectPlayer req = new ReqSelectPlayer();
         req.setPlayerId(playerId);
-        session.sendPacket(req);
+        session.send(req);
     }
 }
