@@ -2,14 +2,12 @@ package org.forfun.mmorpg.framework.eventbus;
 
 /**
  * 事件基类
- * 
- *
  *
  */
 public interface BaseEvent {
 
 	default Object getOwner() {
-		return String.valueOf("system");
+		return "system";
 	}
 
 }
