@@ -1,12 +1,12 @@
 package org.forfun.mmorpg.game;
 
+import jakarta.annotation.PostConstruct;
 import org.forfun.mmorpg.game.base.GameContext;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.core.env.StandardEnvironment;
 import org.springframework.core.io.support.ResourcePropertySource;
 
-import javax.annotation.PostConstruct;
 import java.io.IOException;
 
 @Configuration
@@ -17,12 +17,12 @@ public class ServerConfigFactory {
 
     @PostConstruct
     public void init() throws IOException {
-        environment.getPropertySources().addLast(new ResourcePropertySource("file:config/common.properties"));
         switch (GameContext.serverType) {
             case GATE -> environment.getPropertySources().addLast(new ResourcePropertySource("file:config/gate.properties"));
             case GAME -> environment.getPropertySources().addLast(new ResourcePropertySource("file:config/server.properties"));
             case CENTRE -> environment.getPropertySources().addLast(new ResourcePropertySource("file:config/center.properties"));
         }
+        environment.getPropertySources().addLast(new ResourcePropertySource("file:config/common.properties"));
     }
 
 }
