@@ -1,5 +1,6 @@
 package org.forfun.mmorpg.game.database;
 
+import lombok.extern.slf4j.Slf4j;
 import org.forfun.mmorpg.game.logger.LoggerUtils;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.boot.context.properties.ConfigurationProperties;
@@ -11,6 +12,7 @@ import org.springframework.context.annotation.Primary;
 import javax.sql.DataSource;
 
 @Configuration
+@Slf4j
 public class DataSourcesConfig {
 
 	@Bean(name = "configDataSource")
@@ -18,7 +20,7 @@ public class DataSourcesConfig {
 	@Primary
 	@ConfigurationProperties(prefix = "spring.datasource.configdb")
 	public DataSource primaryDataSource() {
-		LoggerUtils.info("静态数据源建立链接");
+		log.info("静态数据源建立链接");
 		return DataSourceBuilder.create().build();
 	}
 
@@ -26,7 +28,7 @@ public class DataSourcesConfig {
 	@Qualifier("userDataSource")
 	@ConfigurationProperties(prefix = "spring.datasource.userdb")
 	public DataSource secondaryDataSource() {
-		LoggerUtils.info("动态数据源建立链接");
+		log.info("动态数据源建立链接");
 		return DataSourceBuilder.create().build();
 	}
 }
