@@ -1,13 +1,14 @@
 package org.forfun.mmorpg.game.database.config.domain;
 
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.Id;
 import lombok.Getter;
 
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.Id;
+
 
 @Getter
-@Entity
+@Entity(name = "configcommonvalue")
 public class ConfigCommonValue {
 
     @Id
