@@ -1,32 +1,28 @@
 package org.forfun.mmorpg.game.database.converter;
 
-import org.forfun.mmorpg.common.util.JsonUtil;
+import jakarta.persistence.AttributeConverter;
+import jforgame.commons.JsonUtil;
+import org.apache.commons.lang3.StringUtils;
 
-import javax.persistence.AttributeConverter;
 import java.util.Map;
 
-public class JpaMapConverter implements AttributeConverter<Object,String> {
+public class JpaMapConverter implements AttributeConverter<Object, String> {
 
-    /**
-     *
-     * @param attribute
-     * @return
-     */
     @Override
     public String convertToDatabaseColumn(Object attribute) {
         try {
             return JsonUtil.object2String(attribute);
         } catch (Exception e) {
-            return null;
+            throw new RuntimeException(e);
         }
     }
 
     @Override
     public Object convertToEntityAttribute(String dbData) {
-        if(dbData!=null){
+        if (StringUtils.isEmpty(dbData)) {
             return JsonUtil.string2Object(dbData, Map.class);
         }
-        return dbData;
+        return null;
     }
 
 }
