@@ -1,8 +1,9 @@
 package org.forfun.mmorpg.game.database.converter;
 
-import org.forfun.mmorpg.common.util.JsonUtil;
+import jakarta.persistence.AttributeConverter;
+import jforgame.commons.JsonUtil;
+import org.apache.commons.lang3.StringUtils;
 
-import javax.persistence.AttributeConverter;
 import java.util.Set;
 
 public class JpaSetConverter implements AttributeConverter<Object, String> {
@@ -10,15 +11,15 @@ public class JpaSetConverter implements AttributeConverter<Object, String> {
     @Override
     public String convertToDatabaseColumn(Object attribute) {
         try {
-            return JsonUtil.object2String(attribute);
+            return jforgame.commons.JsonUtil.object2String(attribute);
         } catch (Exception e) {
-            return null;
+            throw new RuntimeException(e);
         }
     }
 
     @Override
     public Object convertToEntityAttribute(String dbData) {
-        if (dbData != null) {
+        if (StringUtils.isEmpty(dbData)) {
             return JsonUtil.string2Object(dbData, Set.class);
         }
         return null;
