@@ -1,9 +1,13 @@
 package org.forfun.mmorpg.game.player.facade;
 
+import jforgame.socket.share.IdSession;
+import jforgame.socket.share.annotation.MessageRoute;
+import jforgame.socket.share.annotation.RequestHandler;
 import org.forfun.mmorpg.framework.eventbus.EventBus;
 import org.forfun.mmorpg.game.Modules;
 import org.forfun.mmorpg.game.base.GameContext;
 import org.forfun.mmorpg.game.database.user.entity.PlayerEnt;
+import org.forfun.mmorpg.game.exception.BusinessRequestException;
 import org.forfun.mmorpg.game.gm.GmCommands;
 import org.forfun.mmorpg.game.gm.GmHandler;
 import org.forfun.mmorpg.game.player.event.PlayerLevelUpEvent;
@@ -13,34 +17,36 @@ import org.forfun.mmorpg.game.player.message.ReqPlayerLogin;
 import org.forfun.mmorpg.game.player.message.ReqSelectPlayer;
 import org.forfun.mmorpg.game.player.message.ResPlayerLogin;
 import org.forfun.mmorpg.game.player.service.LoginService;
-import org.forfun.mmorpg.net.socket.IdSession;
-import org.forfun.mmorpg.protocol.annotation.MessageMapping;
-import org.forfun.mmorpg.protocol.annotation.ModuleMeta;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 
 @Controller
-@ModuleMeta(module = Modules.PLAYER)
+@MessageRoute(module = Modules.PLAYER)
 public class PlayerFacade {
 	
 	
 	@Autowired
 	private LoginService loginService;
 
-	@MessageMapping
+	@RequestHandler
 	public void reqAccountLogin(IdSession session, ReqAccountLogin request) {
+		if (request.getAccountId() > 0) {
+			throw new BusinessRequestException(100);
+		}
 		loginService.handleAccountLogin(session, request.getAccountId(), request.getPassword());
 	}
 
+	@RequestHandler
 	public void reqSelectPlayer(IdSession session, ReqSelectPlayer request) {
 		loginService.handleSelectPlayer(session, request.getPlayerId());
 	}
 
+	@RequestHandler
 	public void reqPlayerLogin(IdSession session, ReqPlayerLogin request) {
 		long playerId = request.getPlayerId();
 		System.out.println("角色[" + playerId + "]登录");
 
-		session.sendPacket(new ResPlayerLogin());
+		session.send(new ResPlayerLogin());
 
 		PlayerEnt player = new PlayerEnt();
 		player.setId(playerId);
