@@ -1,10 +1,10 @@
 package org.forfun.mmorpg.game.player.message;
 
+import jforgame.socket.share.annotation.MessageMeta;
+import jforgame.socket.share.message.Message;
 import lombok.Getter;
 import lombok.Setter;
 import org.forfun.mmorpg.game.player.service.PlayerService;
-import org.forfun.mmorpg.protocol.annotation.MessageMeta;
-import org.forfun.mmorpg.protocol.Message;
 
 @MessageMeta(cmd = PlayerService.CMD_REQ_PLAYER_LOGIN)
 @Getter
