@@ -1,11 +1,11 @@
 package org.forfun.mmorpg.game.player.message;
 
 import com.baidu.bjf.remoting.protobuf.annotation.Protobuf;
+import jforgame.socket.share.annotation.MessageMeta;
+import jforgame.socket.share.message.Message;
 import lombok.Data;
 import org.forfun.mmorpg.game.player.message.vo.PlayerLoginVo;
 import org.forfun.mmorpg.game.player.service.PlayerService;
-import org.forfun.mmorpg.protocol.annotation.MessageMeta;
-import org.forfun.mmorpg.protocol.Message;
 
 import java.util.ArrayList;
 import java.util.List;
