@@ -1,5 +1,6 @@
 package org.forfun.mmorpg.game.skill.facade;
 
+import jforgame.socket.share.annotation.MessageRoute;
 import org.forfun.mmorpg.framework.eventbus.Subscribe;
 import org.forfun.mmorpg.game.Modules;
 import org.forfun.mmorpg.game.base.MessageUtil;
@@ -9,7 +10,6 @@ import org.forfun.mmorpg.game.database.user.entity.PlayerEnt;
 import org.forfun.mmorpg.game.player.event.PlayerLevelUpEvent;
 import org.forfun.mmorpg.game.player.event.PlayerLoginEvent;
 import org.forfun.mmorpg.game.skill.message.RespPlayerSkills;
-import org.forfun.mmorpg.protocol.annotation.ModuleMeta;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 
@@ -17,7 +17,7 @@ import java.util.Collection;
 import java.util.Collections;
 
 @Controller
-@ModuleMeta(module = Modules.SKILL)
+@MessageRoute(module = Modules.SKILL)
 public class SkillFacade {
 
 	@Autowired
