@@ -1,7 +1,7 @@
 package org.forfun.mmorpg.test.codec;
 
 
-import org.forfun.mmorpg.net.message.Message;
+import org.forfun.mmorpg.protocol.Message;
 import org.forfun.mmorpg.protocol.annotation.MessageMeta;
 
 @MessageMeta(cmd = 1)
